Corrected a bunch of typos.
I think I corrected all the typos that I found while
working through this tutorial, but please review carefully
because I am new to Angular and could have mis-corrected
one or two.

Lots of typos:
We need some to display on their posts and profile page.
When substituting a customer user model
Some users are admins and can do anywhere,
we want the ability to work will all facets
 we much use
returns makes
Replace Authentication.register when the following:
with using a `ForeignKey`
from authentication.serializers import Account
(should be AccountSerializer in posts/serializers.py)

watches the value of an array from changes

Create static/javascripts/posts/controller/new-post.controller.js
(should be posts/controllerS, not posts/controller)

The worst typo by far, in posts/views.py you have perform_create()
defined at the module level (with tabs, not spaces) rather than
inside of the class PostViewSet.  This is obviously wrong (but took
me hours to find it) because it calls super(PostViewSet, self) thus
it is supposed to be a member of that class.  With the indent wrong,
the javascript returns Error: undefined.